% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{add_audio_js_script}
\alias{add_audio_js_script}
\title{Adds a JavaScript code for zooming of emoji and playing sound}
\usage{
add_audio_js_script()
}
\description{
Adds a JavaScript code for zooming of emoji and playing sound
}
\author{
George Moroz <agricolamz@gmail.com>
}
