% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_annotate.R
\name{color_annotate}
\alias{color_annotate}
\title{Make color}
\usage{
color_annotate(gloss)
}
\arguments{
\item{gloss}{vector of characters.}
}
\description{
Annotate with color.
}
\author{
George Moroz <agricolamz@gmail.com>
}
