% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_examples_db.R
\name{get_examples_db}
\alias{get_examples_db}
\title{Get database of interlinear examples}
\usage{
get_examples_db()
}
\value{
a dataframe with all interlinear examples from rmarkdown document.
}
\description{
Reads database of interlinear examples collected through the whole document.
}
\author{
George Moroz <agricolamz@gmail.com>
}
