% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glosses_df.R
\docType{data}
\name{glosses_df}
\alias{glosses_df}
\title{Catalog of glosses}
\format{
A data frame with 1341 rows and 4 variables:
\describe{
  \item{gloss}{the gloss abbreviation}
  \item{definition_en}{the gloss definition}
  \item{source}{the gloss source. Three possible values: Leipzig Glossing Rules, \href{https://en.wikipedia.org/wiki/List_of_glossing_abbreviations}{Wikipedia} or lingglosses (this means parsed from Glossa).}
  \item{weight}{glossa weight used for the choice in case of multiple definitions per gloss.}
}
}
\usage{
glosses_df
}
\description{
A dataset contains the list of glosses from the Leipzig Glossing Rules by
Comrie, Haspelmath, and Bickel and other glosses automatically gathered from
Glossa Journal articles.
}
\keyword{datasets}
