% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_caps.R
\name{small_caps}
\alias{small_caps}
\title{Make a small caps}
\usage{
small_caps(gloss)
}
\arguments{
\item{gloss}{vector of characters.}
}
\value{
a strings small caps span html tags.
}
\description{
Convert upper script to small caps.
}
\author{
George Moroz <agricolamz@gmail.com>
}
