% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_df.R
\name{convert_to_df}
\alias{convert_to_df}
\title{Converts example to the data.frame and adds it to the database}
\usage{
convert_to_df(
  transliteration,
  glosses,
  free_translation = "",
  comment = "",
  annotation = "",
  drop_transliteration = FALSE,
  write_to_db = TRUE,
  counter = getOption("lingglosses.examples_counter")
)
}
\value{
dataframe with
}
\description{
Converts example to the data.frame and adds it to the database
}
\author{
George Moroz <agricolamz@gmail.com>
}
