% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloss_example.R
\name{gloss_example}
\alias{gloss_example}
\title{Gloss an example}
\usage{
gloss_example(
  transliteration,
  glosses,
  free_translation = "",
  orthography = "",
  comment = "",
  line_length = 70
)
}
\arguments{
\item{transliteration}{blah}

\item{glosses}{blah}

\item{free_translation}{blah}

\item{orthography}{blah}

\item{comment}{blah}

\item{line_length}{blah}
}
\value{
no output
}
\description{
Creates an interlinear glossed example for linguistics.
}
\examples{
gloss_example("bur-e-**ri** c'in-ne-s:u",
               "fly-NPST-**INF** know-HAB-NEG",
               "I cannot fly. (Zilo Andi, East Caucasian)",
               comment = "(lit. do not know how to)")

}
\author{
George Moroz <agricolamz@gmail.com>
}
